dev(ios): fix constraints for iPad
iPad, unlike iPhone, will not use the width of the screen as the width of the popover window for sharing. Instead, it requires the user to provide a location for its popover to originate. This change allows the user to pass in an optional Rect object from Flutter to specify where the popover should locate.